+++ src/app/core/services/constants.service.ts
@@ -1,0 +1,17 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ConstantsService {
+
+  private constValue = 'testConstantsServiceValue';
+
+  constructor() { }
+
+  getConstValue(): any {
+    return this.constValue;
+  }
+}
+
+export const constServiceInstance = new ConstantsService();